new changes : refined property of product, banner event in proper formatted form for frontend convenience, implmented unit of work desing pattern for concurrency control , implmeeted dynamic product pricing on the basis of special event generation, implmented nepal time zone

+++ Application/Dto/BannerEventSpecialDTOs/AddEventRuleDTO.cs
@@ -13,9 +13,9 @@ namespace Application.Dto.BannerEventSpecialDTOs
         public string TargetValue { get; set; } = string.Empty;
         public string? Conditions { get; set; }
         public PromotionType DiscountType { get; set; }
-        public double DiscountValue { get; set; }
-        public double? MaxDiscount { get; set; }
-        public double? MinOrderValue { get; set; }
+        public decimal DiscountValue { get; set; }
+        public decimal? MaxDiscount { get; set; }
+        public decimal? MinOrderValue { get; set; }
         public int Priority { get; set; } = 1;
     }
 }
